+++ packages/apps/desktop-viewer-test/src/frontend/components/routes/ViewerRoute.tsx
@@ -22,6 +22,10 @@ import { useLocation } from "react-router-dom";
 
 import { viewerRpcs } from "../../../common/ViewerConfig";
 import { IModelMergeItemsProvider } from "../../extensions";
+import {
+  getSchemaContext,
+  unifiedSelectionStorage,
+} from "../../selectionStorage";
 
 export interface ViewerRouteState {
   filePath?: string;
@@ -72,6 +76,8 @@ export const ViewerRoute = () => {
         new IModelMergeItemsProvider(),
       ]}
       enablePerformanceMonitors={true}
+      selectionStorage={unifiedSelectionStorage}
+      getSchemaContext={getSchemaContext}
     />
   ) : null;
 };