+++ packages/templates/cra-template-desktop-viewer/template/src/frontend/components/routes/ViewerRoute.tsx
@@ -22,6 +22,7 @@ import { useLocation } from "react-router-dom";
 
 import { viewerRpcs } from "../../../common/ViewerConfig";
 import { IModelMergeItemsProvider } from "../../extensions";
+import { unifiedSelectionStorage, getSchemaContext } from "../../selectionStorage";
 
 export interface ViewerRouteState {
   filePath?: string;
@@ -72,6 +73,8 @@ export const ViewerRoute = () => {
         new IModelMergeItemsProvider(),
       ]}
       enablePerformanceMonitors={true}
+      selectionStorage={unifiedSelectionStorage}
+      getSchemaContext={getSchemaContext}
     />
   ) : null;
 };