+++ packages/templates/cra-template-web-viewer/template/src/App.tsx
@@ -37,6 +37,10 @@ import React, { useCallback, useEffect, useMemo, useState } from "react";
 
 import { Auth } from "./Auth";
 import { history } from "./history";
+import {
+  unifiedSelectionStorage,
+  getSchemaContext,
+} from "./selectionStorage";
 
 const App: React.FC = () => {
   const [iModelId, setIModelId] = useState(process.env.IMJS_IMODEL_ID);
@@ -183,6 +187,8 @@ const App: React.FC = () => {
           }),
           new MeasureToolsUiItemsProvider(),
         ]}
+        selectionStorage={unifiedSelectionStorage}
+        getSchemaContext={getSchemaContext}
       />
     </div>
   );